fix vst playhead issue add ordered stackedUIs for controllers and fastmaps add double click to create node fix undoable coalesced actions remove annoying sounds when pressing keyboard with a keyboard controller

+++ Source/UI/Style.cpp
@@ -234,7 +234,7 @@ void CachedGlyph::updateGlyph(){
         
         glyphArr.clear();
         glyphArr.addCurtailedLineOfText (*getCurFont(), text, 0.0f, 0.0f,
-                                         area.getWidth(), useEllipsesIfTooBig);
+                                         (float)area.getWidth(), useEllipsesIfTooBig);
 
         glyphArr.justifyGlyphs (0, glyphArr.getNumGlyphs(),
                                 area.getX(), area.getY(), area.getWidth(), area.getHeight(),
